// src/lib/api/generated/models/getSubscriptionsParams.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Marketplace API
 * API for Village Website & Marketplace
 * OpenAPI spec version: 1.0.0
 */
import type { ResourceId } from './resourceId';
import type { SubscriptionStatus } from './subscriptionStatus';
import type { UserId } from './userId';

export type GetSubscriptionsParams = {
/**
 * The page number for pagination
 * @minimum 1
 */
page?: number;
/**
 * Number of items to return per page
 * @minimum 1
 * @maximum 50
 */
limit?: number;
/**
 * Filter by buyer ID
 */
buyerId?: UserId & unknown;
/**
 * Filter by seller ID
 */
sellerId?: UserId & unknown;
/**
 * Filter by product ID
 */
productId?: ResourceId & unknown;
/**
 * Filter by subscription status
 */
status?: SubscriptionStatus & unknown;
};